requirements.txt
openai
python-dotenv
newspaper3k
feedparser
lxml_html_clean